// src/types.ts
export type news = {
  session_id: string;
  source_domain: string;
  authors: string;
  date_download: string;
  date_modify: string;
  date_publish: string;
  description: string;
  images: string;
  language: string;
  maintext: string;
  relevance: string;
  sentiment: string;
  summary: string;
  tags: string;
  timestamp: string;
  title: string;
  top_image: string;
  news_url: string;
}